feat: add member directory items
