+++ src/components/members/index.ts
@@ -1,1 +1,3 @@
 export * from './RecentMemberItem';
+export * from './DirectoryMemberItem';
+export * from './MemberCardItem';
